+++ internal/core/software_renderer/fonts.rs
@@ -46,16 +46,16 @@ pub trait GlyphRenderer {
 pub(super) const DEFAULT_FONT_SIZE: LogicalLength = LogicalLength::new(12 as Coord);
 
 mod pixelfont;
-#[cfg(feature = "systemfonts")]
+#[cfg(feature = "software-renderer-systemfonts")]
 pub mod vectorfont;
 
-#[cfg(feature = "systemfonts")]
+#[cfg(feature = "software-renderer-systemfonts")]
 pub mod systemfonts;
 
 #[derive(derive_more::From)]
 pub enum Font {
     PixelFont(pixelfont::PixelFont),
-    #[cfg(feature = "systemfonts")]
+    #[cfg(feature = "software-renderer-systemfonts")]
     VectorFont(vectorfont::VectorFont),
 }
 
@@ -77,7 +77,7 @@ pub fn match_font(request: &FontRequest, scale_factor: ScaleFactor) -> Font {
     let font = match bitmap_font {
         Some(bitmap_font) => bitmap_font,
         None => {
-            #[cfg(feature = "systemfonts")]
+            #[cfg(feature = "software-renderer-systemfonts")]
             if let Some(vectorfont) = systemfonts::match_font(request, scale_factor) {
                 return vectorfont.into();
             }
@@ -86,9 +86,9 @@ pub fn match_font(request: &FontRequest, scale_factor: ScaleFactor) -> Font {
             {
                 fallback_bitmap_font
             } else {
-                #[cfg(feature = "systemfonts")]
+                #[cfg(feature = "software-renderer-systemfonts")]
                 return systemfonts::fallbackfont(request.pixel_size, scale_factor).into();
-                #[cfg(not(feature = "systemfonts"))]
+                #[cfg(not(feature = "software-renderer-systemfonts"))]
                 panic!("No font fallback found. The software renderer requires enabling the `EmbedForSoftwareRenderer` option when compiling slint files.")
             }
         }
@@ -140,7 +140,7 @@ pub fn text_size(
                 max_width.map(|max_width| (max_width.cast() * scale_factor).cast()),
             )
         }
-        #[cfg(feature = "systemfonts")]
+        #[cfg(feature = "software-renderer-systemfonts")]
         Font::VectorFont(vf) => {
             let layout = text_layout_for_font(&vf, &font_request, scale_factor);
             layout.text_size(